Bot Behaviour Changes
Changed how bot will decide which action to do.
Added saving money to buy tower.
Bot will now place towers inside of a smaller box

+++ BTD6-ai/Action.h
@@ -2,7 +2,7 @@
 class Action
 {
 public:
-	bool isReady(double t_currentRound) { return t_currentRound >= round; }
+	bool isReady(int t_currentRound) { return t_currentRound >= round; }
 protected:
 	int round; // Round to do action on
 };
